third_party/blink/web_tests/platform/generic/external/wpt/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_2-expected.txt

This is a testharness.js-based test.
FAIL Removing iframe from document via innerHTML removes it from history assert_equals: History length before iframe removal expected 2 but got 1
Harness: the test ran to completion.
